Refactor(logging): 重构日志模块并统一日志相关类命名
将 app_logger.dart 重命名为 logger.dart，并统一其他日志相关类的命名
添加新的日志组件：log_level.dart, log_entry.dart, log_formatter.dart, log_buffer.dart, file_log_sink.dart
更新所有引用旧日志类名的文件
