+++ lib/app/logging/log_buffer.dart
@@ -1,6 +1,6 @@
 import 'dart:collection';
 
-import 'package:onetj/app/logging/app_log_entry.dart';
+import 'package:onetj/app/logging/log_entry.dart';
 
 class AppLogBuffer {
   AppLogBuffer({required int capacity})
